refactor: move persona-related functions from util.ts to persona.ts
Move getUserName, getUserIcon, getPersonaPrompt, getUserIconProtrait, and checkPersonaBinded functions to persona.ts for better code organization.
