+++ src/ts/persona.ts
@@ -6,6 +6,8 @@ import { language } from "src/lang"
 import { reencodeImage } from "./process/files/inlays"
 import { PngChunk } from "./pngChunk"
 import { v4 } from "uuid"
+import { get } from "svelte/store"
+import { selectedCharID } from "./stores.svelte"
 
 export async function selectUserImg() {
     const selected = await selectSingleFile(['png'])
@@ -146,3 +148,54 @@ export async function importUserPersona() {
         return
     }
 }
+export function getUserName() {
+    const bindedPersona = checkPersonaBinded()
+    if (bindedPersona) {
+        return bindedPersona.name
+    }
+    const db = getDatabase()
+    return db.username ?? 'User'
+}export function checkPersonaBinded() {
+    try {
+        const db = getDatabase()
+        const selectedChar = get(selectedCharID)
+        const character = db.characters[selectedChar]
+        const chat = character.chats[character.chatPage]
+        if (!chat.bindedPersona) {
+            return null
+        }
+        const persona = db.personas.find(v => v.id === chat.bindedPersona)
+        return persona
+    } catch (error) {
+        return null
+    }
+}
+export function getUserIcon() {
+    const bindedPersona = checkPersonaBinded()
+    if (bindedPersona) {
+        return bindedPersona.icon
+    }
+    const db = getDatabase()
+    return db.userIcon ?? ''
+}
+export function getPersonaPrompt() {
+    const bindedPersona = checkPersonaBinded()
+    if (bindedPersona) {
+        return bindedPersona.personaPrompt
+    }
+    const db = getDatabase()
+    return db.personaPrompt ?? ''
+}
+export function getUserIconProtrait() {
+    try {
+        const bindedPersona = checkPersonaBinded()
+        if (bindedPersona) {
+            return bindedPersona.largePortrait
+        }
+        const db = getDatabase()
+        return db.personas[db.selectedPersona].largePortrait
+    } catch (error) {
+        return false
+    }
+}
+
