+++ src/ts/process/embedding/addinfo.ts
@@ -2,7 +2,7 @@ import { getDatabase } from "src/ts/storage/database.svelte";
 import { type Chat } from 'src/ts/storage/types/chat';
 import { type character } from 'src/ts/storage/types/character';
 import { HypaProcesser } from '../memory/hypamemory'
-import { getUserName } from "src/ts/util";
+import { getUserName } from "src/ts/persona";
 
 export async function additionalInformations(char: character,chats:Chat,){
     const processer = new HypaProcesser()
