+++ src/ts/process/memory/supaMemory.ts
@@ -9,7 +9,7 @@ import { HypaProcesser } from "./hypamemory";
 import { stringlizeChat } from "../stringlize";
 import { globalFetch } from "src/ts/globalApi.svelte";
 import { runSummarizer } from "../transformers";
-import { getUserName } from "src/ts/util";
+import { getUserName } from "src/ts/persona";
 import { parseChatML } from "src/ts/parser.svelte";
 
 export async function supaMemory(
