+++ src/ts/process/stringlize.ts
@@ -1,6 +1,6 @@
 import type { OpenAIChat } from "./index.svelte";
 import { getDatabase } from "../storage/database.svelte";
-import { getUserName } from "../util";
+import { getUserName } from "../persona";
 
 export function multiChatReplacer(){
